+++ packages/http-server-mcp/tests/unit/jest.config.ts
@@ -4,10 +4,10 @@ export default {
   ...jestUnitConfig(),
   coverageThreshold: {
     global: {
-      statements: 99.1,
-      branches: 90.7,
-      functions: 99.9,
-      lines: 99.1,
+      statements: 99.28,
+      branches: 92.85,
+      functions: 100,
+      lines: 99.28,
     },
   },
 };